network failure

+++ mobile/src/main/java/net/appmecha/comcast/wear/moviemates/CinemaServices/IFilmServices.java
@@ -7,5 +7,5 @@ import java.util.ArrayList;
  * Created by Neil on 07/06/2015.
  */
 public interface IFilmServices {
-    ArrayList<FilmDetail> GetFilms(long longitude,long latitude);
+    ArrayList<FilmDetail> GetFilms(double longitude,double latitude);
 }
